moved peerscout_api to a seperate folder
